Refactoring: Add TextStream buffer to improve output writing performance
- direct use of std::stringstream and std::ostream gave a 30%
  drop in performance.

+++ src/markdown.cpp
@@ -2729,7 +2729,7 @@ QCString Markdown::process(const QCString &input, int &startNewlines)
   if (p>result.data())
   {
     // strip part of the input
-    result = result.mid(p-result.data());
+    result = result.mid(static_cast<int>(p-result.data()));
   }
   return result;
 }
